Arreglados los botones y una animación en pausa

+++ include/estado_menu.h
@@ -9,8 +9,8 @@ estado_menu.h
 extern void MostrarMenu();
 
 /* Estas funciones dibujan los botones del menú en las coordenadas indicadas */
-extern void dibujar_botonJugar(uint8 X, uint8 Y);
-extern void dibujar_botonContinuar(uint8 X, uint8 Y);
-extern void dibujar_botonSalir(uint8 X, uint8 Y);
+extern void dibujar_botonJugar(int X, int Y);
+extern void dibujar_botonContinuar(int X, int Y);
+extern void dibujar_botonSalir(int X, int Y);
 
 #endif // E_MENU_H
